Major changes to venue manager with summary totals at the top, as well as  totals per product row. Also, needed to add ajax functionality for making a payment.  The code for a venue column in the product listing has been written but is currently not implemented based on lack of space in display.

+++ includes/activation-deactivation.php
@@ -25,7 +25,7 @@ function taste_add_venue_table() {
 			voucher_pct FLOAT,
 			paid_member TINYINT(1) ZEROFILL NOT NULL DEFAULT 0, 
 			member_renewal_date DATETIME,
-			membership_cost FLOAT,
+			membership_cost DECIMAL(10,2),
 			PRIMARY KEY (venue_id),
 			UNIQUE KEY (name),
 			KEY (venue_type)
